nf bug: sx driver dummy, irq should be u16

+++ mLRS/Common/sx-drivers/sx12xx_driver.h
@@ -53,7 +53,7 @@ class SxDriverDummy
 
     void SetRfPower_dbm(int8_t power_dbm) {}
     void SetLoraConfigurationByIndex(uint8_t index) {}
-    void ClearIrqStatus(uint8_t IrqMask) {}
+    void ClearIrqStatus(uint16_t IrqMask) {}
 
     int16_t ReceiverSensitivity_dbm(void) { return 0; }
     int8_t RfPower_dbm(void) { return INT8_MIN; }
